# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Karigar Service Marketplace API",
        "description": "RESTful API for Karigar - A service marketplace connecting customers with skilled service providers (electricians, plumbers, carpenters, etc.)",
        "contact": {
            "name": "Karigar API Support",
            "email": "support@karigar.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Karigar API Server"
        }
    ],
    "paths": {
...
        "/api/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get user profile",
                "description": "Retrieve current authenticated user's profile information",
                "operationId": "ede45b19680d6fb1319013a741a05734",
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profile retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "user@example.com"
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Doe"
                                                },
                                                "prefix": {
                                                    "type": "string",
                                                    "example": "Mr."
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "example": "johndoe"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+1234567890"
                                                },
                                                "alternate_phone": {
                                                    "type": "string",
                                                    "example": "+0987654321"
                                                },
                                                "profile_picture": {
                                                    "type": "string",
                                                    "example": "https://api.example.com/storage/profiles/user-1.jpg"
                                                },
                                                "provider_profile": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "category": {
                                                            "type": "string",
                                                            "example": "Plumber"
                                                        },
                                                        "location": {
                                                            "type": "string"
                                                        },
                                                        "is_approved": {
                                                            "type": "boolean"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "roles": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "viewer"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/update": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update profile information",
                "description": "Update user's personal profile details (name, email, phone, etc.)",
                "operationId": "658523cb93a6af7d87e1ca8740b514f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "description": "First name (max 255 chars)",
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "description": "Last name (max 255 chars)",
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "description": "Email address (must be unique)",
                                        "type": "string",
                                        "format": "email",
                                        "example": "newemail@example.com"
                                    },
                                    "prefix": {
                                        "description": "Name prefix (Mr., Mrs., Dr., etc.)",
                                        "type": "string",
                                        "example": "Mr."
                                    },
                                    "username": {
                                        "description": "Username (max 255 chars)",
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "phone": {
                                        "description": "Primary phone number",
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "alternate_phone": {
                                        "description": "Alternate phone number",
                                        "type": "string",
                                        "example": "+0987654321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/update-picture": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Upload profile picture",
                "description": "Upload or update user's profile picture (JPEG, PNG, GIF - max 5MB)",
                "operationId": "9be0f467901353522a713343485bfe99",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "picture"
                                ],
                                "properties": {
                                    "picture": {
                                        "description": "Profile picture file (JPEG, PNG, GIF - max 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile picture uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profile picture updated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "profile_picture": {
                                                    "type": "string",
                                                    "example": "profiles/user-1-abc123.jpg"
                                                },
                                                "profile_picture_url": {
                                                    "type": "string",
                                                    "example": "https://api.example.com/storage/profiles/user-1-abc123.jpg"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid file or file too large"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}